requirements.txt
Flask==0.10
Pillow==2.1.0